Finalise and refactor calculate functionality
Completely refactored the operate function in order to, hopefully, improve
readability and ease of use when adding new functionality or refactoring
old functionality.

Separated display content updates to a new function and changed the way
first and second operands are updated. This at least somewhat fixes the
problem of having too many ifs with confusing conditionals.