Remove extra operators and resize calculator
Removed a few of the operator buttons that were built into the HTML at
the start of the project, but hadn´t been wired up to any javascript yet.
Buttons won´t be adding anything in terms of challenges when it comes to
building/refactoring logic, so in the grand scheme of things they seem
unnecessary.

The idea of this project was to get practice in HTML, javascript and CSS.
So more operators for the sake of having more seems pointless.