Create updateHistory functionality
Wrote first version of the history function. The idea is to have a function
which creates a new element containing the data and result of a calculation.
This element is then appended to the history container to a maximum of 10,
at which point the last child element will be removed and the new element
will be inserted at the top again.