Finalise history functionality
Finished the history functionality. Also added a rounding method on all
results in order to prevent cluttering of the screen.

In a calculator for commercial use I would have opted for a more accurate
solution of course, however since time is limited it's best to move on to
new projects.